implement SIP conversion
add classes in `warpfield/distortion.py`.

The class `Sip` converts input coordinates with a function defined by the
SIP convention. The method `apply` is used to convert _distorted_ coordinates
into _correct_ coordinates.

The class `SipMod` has an additional parameter `center`, which defines the
center of the distortion. The distortion center positions are displaced by
the `center`. Thus, the conversion is no more along with the SIP convention.

The classes `SipDistortion` and `SipModDistortion` work as distortion functions
in the context of `warpfield`. Thus, the instances of the two classes are
executable. They convert _correct_ coordinates into _distorted_ coordinates,
providing an inverse function to the `apply` method.